refactor(apim): refactor model and utils
