+++ packages/fx-core/src/plugins/resource/apim/index.ts
@@ -2,13 +2,12 @@
 // Licensed under the MIT license.
 import { Plugin, FxError, PluginContext, SystemError, UserError, Result, err, ok, QTreeNode, Stage, Func } from "@microsoft/teamsfx-api";
 import { BuildError, UnhandledError } from "./error";
-import { Telemetry } from "./telemetry";
+import { Telemetry } from "./utils/telemetry";
 import { AadPluginConfig, ApimPluginConfig, FunctionPluginConfig, SolutionConfig } from "./model/config";
-import { AadDefaultValues, PluginLifeCycle, PluginLifeCycleToProgressStep, ProgressMessages, ProgressStep, ProjectConstants } from "./constants";
+import { AadDefaultValues, PluginLifeCycle, PluginLifeCycleToProgressStep, ProgressMessages, ProgressStep, ProjectConstants, OperationStatus } from "./constants";
 import { Factory } from "./factory";
-import { ProgressBar } from "./util/progressBar";
+import { ProgressBar } from "./utils/progressBar";
 import { buildAnswer } from "./model/answer";
-import { OperationStatus } from "./model/operation";
 
 export class ApimPlugin implements Plugin {
     private progressBar: ProgressBar = new ProgressBar();
