+++ packages/fx-core/src/plugins/resource/apim/manager/apimManager.ts
@@ -4,11 +4,11 @@ import { ApimDefaultValues, ApimPluginConfigKeys, TeamsToolkitComponent } from "
 import { AssertConfigNotEmpty, AssertNotEmpty } from "../error";
 import { IAadPluginConfig, IApimPluginConfig, IFunctionPluginConfig, ISolutionConfig } from "../model/config";
 import { ApimService } from "../service/apimService";
-import { OpenApiProcessor } from "../util/openApiProcessor";
+import { OpenApiProcessor } from "../utils/openApiProcessor";
 import { IAnswer } from "../model/answer";
 import { LogProvider, TelemetryReporter } from "@microsoft/teamsfx-api";
-import { Lazy } from "../util/lazy";
-import { NamingRules } from "../util/namingRules";
+import { Lazy } from "../utils/commonUtils";
+import { NamingRules } from "../utils/namingRules";
 
 export class ApimManager {
     private readonly logger?: LogProvider;
