+++ packages/fx-core/src/plugins/resource/apim/manager/scaffoldManager.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 import { ApimDefaultValues, ProjectConstants } from "../constants";
-import { OpenApiProcessor } from "../util/openApiProcessor";
+import { OpenApiProcessor } from "../utils/openApiProcessor";
 import { LogProvider, TelemetryReporter } from "@microsoft/teamsfx-api";
 import * as path from "path";
 import * as fs from "fs-extra";
