+++ packages/fx-core/src/plugins/resource/apim/model/aadResource.ts
