+++ packages/fx-core/src/plugins/resource/apim/model/apimResource.ts
@@ -4,10 +4,3 @@ export interface IApimServiceResource {
     serviceName: string;
     resourceGroupName: string;
 }
-
-export interface IAadServiceResource {
-    clientId: string;
-    objectId: string;
-    clientSecretId: string;
-    clientSecret?: string;
-}
