+++ packages/fx-core/src/plugins/resource/apim/model/errorHandlerResult.ts
@@ -1,3 +1,5 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
 export enum ErrorHandlerResult {
     Continue = "Continue",
     Return = "Return",
