+++ packages/fx-core/src/plugins/resource/apim/service/aadService.ts
@@ -2,13 +2,13 @@
 // Licensed under the MIT license.
 import { AadOperationError, AssertNotEmpty, BuildError } from "../error";
 import { AxiosInstance, AxiosResponse, Method } from "axios";
-import { IAadInfo, IPasswordCredential, IServicePrincipal, IServicePrincipals } from "../model/aadResponse";
+import { IAadInfo, IPasswordCredential, IServicePrincipal, IServicePrincipals } from "../model/aadResource";
 import { ErrorHandlerResult } from "../model/errorHandlerResult";
-import { AzureResource, IName, OperationStatus, Operation } from "../model/operation";
+import { AzureResource, IName, OperationStatus, Operation } from "../constants";
 import { LogProvider, TelemetryReporter } from "@microsoft/teamsfx-api";
-import { LogMessages } from "../log";
-import { Telemetry } from "../telemetry";
-import { RetryHandler } from "../util/retryHandler";
+import { LogMessages } from "../utils/log";
+import { Telemetry } from "../utils/telemetry";
+import { RetryHandler } from "../utils/commonUtils";
 
 export class AadService {
     private readonly logger?: LogProvider;
