+++ packages/fx-core/src/plugins/resource/apim/utils/log.ts
@@ -1,8 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
-import { ApimDefaultValues, ProjectConstants } from "./constants";
-import { IName } from "./model/operation";
-import { capitalizeFirstLetter } from "./util";
+import { ApimDefaultValues, ProjectConstants, IName } from "../constants";
+import { capitalizeFirstLetter } from "./commonUtils";
 
 export class LogMessages {
     public static readonly operationStarts = (operation: IName, resourceType: IName, resourceId?: string): string =>
