+++ packages/fx-core/src/plugins/resource/apim/utils/namingRules.ts
