+++ packages/fx-core/src/plugins/resource/apim/utils/progressBar.ts
