+++ packages/fx-core/tests/plugins/resource/apim/namingRules.test.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 import "mocha";
 import chai from "chai";
-import { NamingRules } from "../../../../src/plugins/resource/apim/util/namingRules";
+import { NamingRules } from "../../../../src/plugins/resource/apim/utils/namingRules";
 
 describe("NamingRules", () => {
     describe("#sanitizeId()", () => {
